requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
opencv-python>=4.6.0
pillow>=9.0.0
matplotlib>=3.5.0
pytest>=7.0.0
pyyaml>=6.0
